% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_test_OGS.R
\name{theme_test_OGS}
\alias{theme_test_OGS}
\title{Prism-esque Plots}
\usage{
theme_test_OGS()
}
\description{
This is just a test version of a theme, looks like Prism plots.
}
\details{
Properties as follows:\preformatted{  legend.position = "none"

  plot.title = element_text(hjust = 0.5, vjust = 4, face = "bold", size = 16)

  axis.line = element_line(size = 1.4)

  axis.text.x = element_text(face = "italic", size = 14)

  axis.text.y = element_text(size = 12)

  plot.margin = unit(c(1,2,1,2),"cm")
}
}
\examples{
"plot <- ggplot(x,y,data) + theme_test_OGS()"

}
